+++ lib/hooks/useSupabaseAuth.ts
@@ -39,13 +39,14 @@ const useSupabaseAuth = (loadSession = false) => {
   }, []);
 
   return {
-    signIn: (data: SignInWithOAuthCredentials) =>
+    signIn: (data: SignInWithOAuthCredentials) => {
       supabase.auth.signInWithOAuth({
         ...data,
         options: data.options ?? {
           redirectTo: process.env.NEXT_PUBLIC_BASE_URL ?? "/",
         },
-      }),
+      });
+    },
     signOut: () => supabase.auth.signOut(),
     user,
     sessionToken,